# dashboard/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Numeraire SwissKnife REST API",
    "description": "This API is available to anyone with a Numeraire account. The `Me` (`/me`) endpoints expose the authenticated account and its wallet-scoped operations.",
    "contact": {
      "name": "Dario Anongba Varela",
      "email": "dario.varela@numeraire.tech"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/me/wallets/{wallet_id}/payments": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "List payments for a wallet.",
        "operationId": "list_wallet_payments",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Total amount of results to return",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset where to start returning results",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "ids",
            "in": "query",
            "description": "List of IDs",
            "required": false,
            "schema": {
              "type": [
                "array",
                "null"
              ],
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "wallet_id",
            "in": "query",
            "description": "Wallet ID. Automatically populated with your ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/PaymentStatus"
                }
              ]
            }
          },
          {
            "name": "ledger",
            "in": "query",
            "description": "Ledger",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/Ledger"
                }
              ]
            }
          },
          {
            "name": "ln_addresses",
            "in": "query",
            "description": "Lightning addresses",
            "required": false,
            "schema": {
              "type": [
                "array",
                "null"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "btc_addresses",
            "in": "query",
            "description": "Bitcoin addresses",
            "required": false,
            "schema": {
              "type": [
                "array",
                "null"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "order_direction",
            "in": "query",
            "description": "Direction of the ordering of results",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          },
          {
            "name": "wallet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Send a payment from a wallet.",
        "operationId": "wallet_pay",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payment Sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"422 Unprocessable Entity\",\n    \"reason\": \"Validation failed: ...\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Me"
        ],
        "summary": "Delete failed payments for a wallet.",
        "operationId": "delete_failed_payments",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      }
    },
    "/v1/me/wallets/{wallet_id}/payments/{id}": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Get a wallet payment.",
        "operationId": "get_wallet_payment",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      }
    },
...
    "/v1/payments/{id}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Find a payment",
        "description": "Returns the payment by its ID.",
        "operationId": "get_payment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Payments"
        ],
        "summary": "Delete a payment",
        "description": "Deletes a payment by ID. Returns an empty body. Deleting a payment can affect the wallet balance.",
        "operationId": "delete_payment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"400 Bad Request\",\n    \"reason\": \"Missing required parameter in request\"\n}\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"401 Unauthorized\",\n    \"reason\": \"Invalid credentials\"\n}\n"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"403 Forbidden\",\n    \"reason\": \"Missing permissions\"\n}\n"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"404 Not Found\",\n    \"reason\": \"Resource not found\"\n}\n"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": "\n{\n    \"status\": \"500 Internal Server Error\",\n    \"reason\": \"Internal server error, Please contact your administrator or try later\"\n}\n"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BtcPayment": {
        "type": "object",
        "description": "On-chain Bitcoin details of a payment.",
        "required": [
          "address",
          "txid"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Destination Bitcoin address. Populated for Bitcoin onchain payments."
          },
          "block_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Bitcoin block height where the transaction was confirmed.",
            "minimum": 0
          },
          "txid": {
            "type": "string",
            "description": "Transaction ID for on-chain payments."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Application error response",
        "required": [
          "status",
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Error reason",
            "example": "error message"
          },
          "status": {
            "type": "string",
            "description": "Error status",
            "example": "401 Unauthorized"
          }
        }
      },
...
      "Ledger": {
        "type": "string",
        "description": "The ledger a transaction settles on.",
        "enum": [
          "Lightning",
          "Internal",
          "Onchain"
        ]
      },
...
      "OrderDirection": {
        "type": "string",
        "description": "Direction of result ordering for list endpoints.",
        "enum": [
          "Asc",
          "Desc"
        ]
      },
      "Payment": {
        "type": "object",
        "description": "An outgoing payment, over Lightning, on-chain, or internal to the instance.",
        "required": [
          "id",
          "wallet_id",
          "amount_msat",
          "ledger",
          "status",
          "created_at"
        ],
        "properties": {
          "amount_msat": {
            "type": "integer",
            "format": "int64",
            "description": "Amount in millisatoshis.",
            "minimum": 0
          },
          "bitcoin": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BtcPayment",
                "description": "Bitcoin on-chain payment details"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date of creation in database"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Description"
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message",
            "example": "failed to pay error message"
          },
          "fee_msat": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Fees paid. Populated when a new channel is opened to receive the funds",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Internal ID"
          },
          "internal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InternalPayment",
                "description": "Internal payment details"
              }
            ]
          },
          "ledger": {
            "$ref": "#/components/schemas/Ledger",
            "description": "Ledger"
          },
          "lightning": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LnPayment",
                "description": "Lightning payment details"
              }
            ]
          },
          "payment_time": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Payment time"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus",
            "description": "Status"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Date of update in database"
          },
          "wallet_id": {
            "type": "string",
            "format": "uuid",
            "description": "Wallet ID"
          }
        }
      },
      "PaymentStatus": {
        "type": "string",
        "description": "Lifecycle status of a payment.",
        "enum": [
          "Pending",
          "Settled",
          "Failed"
        ]
      },
...
    },
...
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
...
}